Add file.lastpath to fix regression in projects using malloc://

+++ shlr/sdb/src/cdb.c
@@ -101,8 +101,8 @@ bool cdb_read(struct cdb *c, char *buf, ut32 len, ut32 pos) {
 		return false;
 	}
 	while (len > 0) {
-		ssize_t r = read (c->fd, buf, len);
-		if (r < 1 || (ut32) r != len) {
+		int r = (int)read (c->fd, buf, len);
+		if (r < 1 || (ut32)r != len) {
 			return false;
 		}
 		buf += r;
